Time: 4 ms (84.52%), Space: 17.6 MB (26.93%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -19,7 +19,7 @@ public:
         while(!q.empty())
         {
             int size = q.size();
-            int min = q.front().second;
+            long long int min = q.front().second;
            
             int first,last;
             for(int i=0; i<size; i++)
